feat: total refactor
