+++ main.go
@@ -5,16 +5,27 @@ import (
 	"log"
 	"strconv"
 
+	"github.com/habibiefaried/dns-over-tor-resolver/cachehandler"
+	"github.com/habibiefaried/dns-over-tor-resolver/config"
 	"github.com/habibiefaried/dns-over-tor-resolver/resolvehandler"
 	"github.com/miekg/dns"
 )
 
 func main() {
 	var torResolve *resolvehandler.TorResolve
-	resolverbesidetor := getAllBesideTORResolver()
+	caches, err := cachehandler.InitCachingSystem()
+	if err != nil {
+		log.Fatal(err)
+	}
 
+	c, err := config.ReadConfig()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	resolverbesidetor := resolvehandler.GetAllBesideTORResolver(c, *caches)
 	go func() {
-		torResolve = getTORResolver()
+		torResolve = resolvehandler.GetTORResolver(c, *caches)
 	}()
 
 	// attach request handler func
@@ -77,7 +88,7 @@ func main() {
 	server := &dns.Server{Addr: ":" + strconv.Itoa(port), Net: "udp"}
 	log.Printf("Listening at %d/udp\n", port)
 
-	err := server.ListenAndServe()
+	err = server.ListenAndServe()
 	defer server.Shutdown()
 	if err != nil {
 		log.Fatalf("Failed to start server: %s\n ", err.Error())
